requirements.txt
Flask
flask-cors
google-generativeai
requests
python-dotenv
gunicorn
